requirements.txt
numpy==1.24.3
torch==2.1.2
tqdm==4.65.0
PyYAML==6.0
tensorboard==2.13.0
opencv-python==4.7.0.72
gymnasium==0.28.1
gymnasium[atari]==0.28.1
gymnasium[accept-rom-license]==0.28.1
matplotlib==3.7.1
ale-py==0.8.1
autorom[accept-rom-license]==0.6.1